+++ src/ppidd_prototypes.h
@@ -41,12 +41,12 @@
 
  extern int  PPIDD_Eaf_open (char *name, int type, int *handle);
  extern int  PPIDD_Eaf_write (int handle, double byte_offset, void *buff, int64_t byte_length);
- extern int  PPIDD_Eaf_awrite (int handle, double byte_offset, void *buff, int64_t byte_length, int64_t *request_id);
+ extern int  PPIDD_Eaf_awrite (int handle, double byte_offset, void *buff, int64_t byte_length, int *request_id);
  extern int  PPIDD_Eaf_read (int handle, double byte_offset, void *buff, int64_t byte_length);
- extern int  PPIDD_Eaf_aread (int handle, double byte_offset, void *buff, int64_t byte_length, int64_t *request_id);
- extern int  PPIDD_Eaf_wait (int handle, int64_t request_id);
+ extern int  PPIDD_Eaf_aread (int handle, double byte_offset, void *buff, int64_t byte_length, int *request_id);
+ extern int  PPIDD_Eaf_wait (int handle, int request_id);
  extern int  PPIDD_Eaf_waitall (int64_t *list, int num);
- extern int  PPIDD_Eaf_probe (int64_t request_id, int *status);
+ extern int  PPIDD_Eaf_probe (int request_id, int *status);
  extern int  PPIDD_Eaf_close (int handle);
  extern int  PPIDD_Eaf_delete (char *name);
  extern int  PPIDD_Eaf_length (int handle, double *fsize);
